Remove references to the `light-level` CSS media feature.
It was removed from the Media Queries Level 5 spec in
w3c/csswg-drafts@f5b663c27d5a in favor of the `prefers-contrast` and
`prefers-color-scheme` media features. Reword the paragraph to mention those
instead.

While here, change the biblio for the `matchMedia` API from CSSOM to
CSSOM-VIEW-1, which is where it is actually specified.